feat(MAM-10): 레스토랑 상세 페이지에 위치 섹션 컴포넌트 추가 및 기존 코드 정리
